Crate line_index

See LineIndex.

Structs

LineCol

(line, column) information in the native, UTF-8 encoding.

LineIndex

Maps flat TextSize offsets to/from (line, column) representation.

TextRange

A range in text, represented as a pair of TextSize.

TextSize

A measure of text length. Also, equivalently, an index into text.

WideChar 🔒

WideLineCol

(line, column) information in wide encodings.

Enums

WideEncoding

A kind of wide character encoding.

Functions

analyze_source_file 🔒

This is adapted from the rustc_span crate, https://github.com/rust-lang/rust/blob/de59844c98f7925242a798a72c59dc3610dd0e2c/compiler/rustc_span/src/analyze_source_file.rs

analyze_source_file_dispatch 🔒

analyze_source_file_generic 🔒

analyze_source_file_sse2 🔒 ⚠

Checks 16 byte chunks of text at a time. If the chunk contains something other than printable ASCII characters and newlines, the function falls back to the generic implementation. Otherwise it uses SSE2 intrinsics to quickly find all newlines.