line_index

Struct WideChar

pub(crate) struct WideChar {
    pub(crate) start: TextSize,
    pub(crate) end: TextSize,
}

Fields
start: TextSize

Start offset of a character inside a line, zero-based.

end: TextSize

End offset of a character inside a line, zero-based.

Implementations

impl WideChar

pub(crate) fn len(&self) -> TextSize

Returns the length in 8-bit UTF-8 code units.

pub(crate) fn wide_len(&self, enc: WideEncoding) -> u32

Returns the length in UTF-16 or UTF-32 code units.